clippy + fmt

+++ crates/brontes-classifier/src/discovery_only.rs
@@ -90,7 +90,7 @@ impl<'db, T: TracingProvider, DB: LibmdbxReader + DBWriter> DiscoveryOnlyClassif
                             gas_used:            trace.gas_used,
                             effective_gas_price: trace.effective_price,
                             priority_fee:        trace.effective_price
-                                - (header.base_fee_per_gas.unwrap_or_default() as u128),
+                                - header.base_fee_per_gas.unwrap_or_default(),
                         },
                         data_store: NodeData(vec![Some(action)]),
                     };
